add 戰鬥房間邏輯（TODO：卡牌使用、消息接口接受處理，批消息處理） TODO：rabbitmq

+++ base-core/src/main/java/core/dto/acct/dto/UserInfo.java
@@ -22,7 +22,7 @@ public class UserInfo {
 
     private Long accountId;
 
-    private Long area;
+    private Byte area;
 
     public String getEmail() {
         return email;
@@ -81,11 +81,11 @@ public class UserInfo {
         this.accountId = accountId;
     }
 
-    public Long getArea() {
+    public Byte getArea() {
         return area;
     }
 
-    public void setArea(Long area) {
+    public void setArea(Byte area) {
         this.area = area;
     }
 }
